Extract SpecDescriptor#runSpec
