+++ prototypes/jupiter-engine/src/main/java/info/javaspec/engine/Executable.java
@@ -1,0 +1,6 @@
+package info.javaspec.engine;
+
+@FunctionalInterface
+public interface Executable {
+  void execute();
+}
