+++ prototypes/jupiter-engine/src/main/java/info/javaspec/engine/SpecTestEngine.java
@@ -1,14 +1,13 @@
 package info.javaspec.engine;
 
 import org.junit.platform.engine.*;
-import org.junit.platform.engine.support.descriptor.AbstractTestDescriptor;
 import org.junit.platform.engine.support.descriptor.EngineDescriptor;
 
 public class SpecTestEngine implements TestEngine {
   @Override
   public TestDescriptor discover(EngineDiscoveryRequest discoveryRequest, UniqueId engineId) {
     EngineDescriptor engineDescriptor = new EngineDescriptor(engineId, "JavaSpec");
-    TestDescriptor specDescriptor = SpecDescriptor.forSpec(engineId, "greets the world");
+    TestDescriptor specDescriptor = SpecDescriptor.forSpec(engineId, "greets the world", () -> GreeterSpecs.incrementRunCount());
     engineDescriptor.addChild(specDescriptor);
     return engineDescriptor;
   }
@@ -21,7 +20,8 @@ public class SpecTestEngine implements TestEngine {
 
     for (TestDescriptor childDescriptor : engineDescriptor.getChildren()) {
       listener.executionStarted(childDescriptor);
-      new GreeterSpecs().getOnlySpec().run();
+      SpecDescriptor specDescriptor = (SpecDescriptor) childDescriptor;
+      specDescriptor.runSpec();
       listener.executionFinished(childDescriptor, TestExecutionResult.successful());
     }
 
@@ -31,23 +31,5 @@ public class SpecTestEngine implements TestEngine {
   @Override
   public String getId() {
     return "javaspec-engine";
-  }
-
-  private static class SpecDescriptor extends AbstractTestDescriptor {
-    public static SpecDescriptor forSpec(UniqueId engineId, String behavior) {
-      return new SpecDescriptor(
-        engineId.append("spec", behavior),
-        behavior
-      );
-    }
-
-    private SpecDescriptor(UniqueId uniqueId, String displayName) {
-      super(uniqueId, displayName);
-    }
-
-    @Override
-    public Type getType() {
-      return Type.TEST;
-    }
   }
 }
